Kontroler turnieju ukończony. Rest API gotowe. Tworzenie nie działa z zagnieżdżonymi elementami

+++ src/main/java/Przyrost3/services/FightingSchoolServiceImpl.java
@@ -48,7 +48,7 @@ public class FightingSchoolServiceImpl implements FightingSchoolService {
 
     @Override
     public FightingSchool getBest() {
-        double max = 0;
+        int max = 0;
         int maxId = 0;
         Iterable<FightingSchool> list = repository.findAll();
         for (FightingSchool obj : list) {
